Refactor garbage collector to finalizer
